refactor stm32 UART code
